Separate native_sub_window* into new library
... in order to decouple and remove the circular dependency
between the vk server and the main host server.

Bug: b/537737772
Test: bazel test //host/...
Low-Coverage-Reason: REFACTOR_ONLY
