+++ host/CMakeLists.txt
@@ -52,6 +52,7 @@ add_subdirectory(tracing)
 add_subdirectory(decoder_common)
 add_subdirectory(compressed_textures)
 add_subdirectory(renderdoc)
+add_subdirectory(native_window)
 
 add_library(
     gfxstream_backend.headers
@@ -102,15 +103,7 @@ set(stream-server-core-sources
     virtio_gpu_ring_blob.cpp
     virtio_gpu_timelines.cpp
     vsync_thread.cpp)
-if (APPLE)
-    set(stream-server-core-platform-sources native_sub_window_cocoa.mm)
-elseif (WIN32)
-    set(stream-server-core-platform-sources native_sub_window_win32.cpp)
-elseif (QNX)
-    set(stream-server-core-platform-sources native_sub_window_qnx.cpp)
-else()
-    set(stream-server-core-platform-sources native_sub_window_x11.cpp)
-endif()
+
 
 
 
@@ -120,7 +113,6 @@ add_library(
         gfxstream_backend_static
         STATIC
         ${stream-server-core-sources}
-        ${stream-server-core-platform-sources}
 )
 target_link_libraries(
         gfxstream_backend_static
@@ -146,6 +138,7 @@ target_link_libraries(
         gfxstream-vulkan-server
         GLES_CM_translator_static
         renderControl_dec
+        gfxstream_host_native_window
         )
 if(NOT APPLE AND NOT WIN32)
     target_link_libraries(gfxstream_backend_static PUBLIC gfxstream_xcb_headers)
