+++ host/vulkan/BUILD.bazel
@@ -164,6 +164,7 @@ cc_library(
         "//host/gl/OpenGLESDispatch:gfxstream_host_openglesdispatch_headers",
         "//host/iostream:gfxstream_host_iostream",
         "//host/library:gfxstream_host_library",
+        "//host/native_window:gfxstream_host_native_window",
         "//host/renderdoc:gfxstream_host_renderdoc",
         "//host/tracing:gfxstream_host_tracing",
         "//host/vulkan/cereal:gfxstream_vulkan_cereal",
