+++ host/vulkan/meson.build
@@ -73,6 +73,7 @@ lib_vulkan_server = static_library(
     inc_cereal_common,
     inc_cereal,
     inc_common_base,
+    inc_common_logging,
     inc_common_utils,
     inc_gl_openglesdispatch,
     inc_gl_server,
@@ -84,7 +85,7 @@ lib_vulkan_server = static_library(
     inc_host_include,
     inc_host_iostream,
     inc_host_library,
-    inc_common_logging,
+    inc_host_native_window,
     inc_host_renderdoc,
     inc_host_snapshot,
     inc_host_tracing,
@@ -104,6 +105,7 @@ lib_vulkan_server = static_library(
     lib_host_compressed_textures,
     lib_host_features,
     lib_host_library,
+    lib_host_native_window,
     lib_host_tracing,
     lib_vulkan_cereal,
   ],
